_opentelemetry-py/requirements.txt
pymongo==3.12.0
Flask==3.0.2
requests==2.31.0
opentelemetry-api>=1.21.0
opentelemetry-distro==0.44b0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http>=0.42b0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-flask>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-instrumentation-wsgi==0.44b0
opentelemetry-instrumentation-pymongo>=0.42b0
opentelemetry-proto==1.23.0
opentelemetry-sdk>=1.21.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
psutil==5.8.0